refactor(integrations/airtable): consolidate action files into record and table modules
